refactoring CatalogViewController
